pyproject.toml
[project]
name = "ollama_apps"
version = "0.1.0"
description = "ollama apps. python based apps."
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.9, <=3.13"
authors = [{ name = "chaitu-ycr", email = "chaitu.ycr@gmail.com" }]
keywords = ["python", "ollama", "experiments"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows"
]
dependencies = [
    "beautifulsoup4",
    "chromadb",
    "langchain",
    "langchain-community",
    "ollama",
    "openpyxl",
    "pandas",
    "pypdf",
    "python-docx",
    "python-dotenv",
    "python-pptx",
    "requests",
    "sentence-transformers",
    "streamlit",
]

[project.optional-dependencies]
dev = [
    "mkdocs-material",
    "mkdocstrings-python",
    "mkdocs-macros-plugin",
    "mkdocs-include-markdown-plugin",
]

[project.urls]
homepage = "https://github.com/chaitu-ycr/ollama_apps"
repository = "https://github.com/chaitu-ycr/ollama_apps"
documentation = "https://chaitu-ycr.github.io/ollama_apps/"

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"
